# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateSubmissionInput:
    properties:
      answer_text:
        type: string
      file_url:
        type: string
    required:
    - answer_text
    type: object
...
  dto.ReviewSubmissionInput:
    properties:
      score:
        type: integer
      status:
        type: string
    required:
    - score
    - status
    type: object
...
  dto.UpdateSubmissionInput:
    properties:
      answer_text:
        type: string
      file_url:
        type: string
    required:
    - answer_text
    type: object
info:
  contact: {}
  description: Backend API for Mini-LMS / Teacher Platform
  title: Teacher Platform API
  version: "1.0"
paths:
...
  /assignments/{id}/submissions:
    get:
      consumes:
      - application/json
      description: Teacher/Admin gets all submissions for an assignment
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get submissions by assignment
      tags:
      - submissions
    post:
      consumes:
      - application/json
      description: Student submits a solution for an assignment
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create submission input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubmissionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create submission
      tags:
      - submissions
...
  /submissions/{id}:
    delete:
      consumes:
      - application/json
      description: Student deletes own submission before deadline
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete submission
      tags:
      - submissions
    get:
      consumes:
      - application/json
      description: Get one submission by id
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get submission by id
      tags:
      - submissions
    patch:
      consumes:
      - application/json
      description: Student updates own submission before deadline
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update submission input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubmissionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update submission
      tags:
      - submissions
  /submissions/{id}/review:
    patch:
      consumes:
      - application/json
      description: Teacher/Admin reviews a submission and sets score/status
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review submission input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewSubmissionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Review submission
      tags:
      - submissions
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"